microservices/monitor-sensors-service/pyproject.toml
[tool.poetry]
name = "monitor-sensors-service"
version = "0.1.0"
description = "Watchit main microservice for monitoring sensors data"
authors = ["Liad Kashanovsky"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
astroid = "3.1.0"
async-timeout = "4.0.3"
asyncio = "3.4.3"
attrs = "23.2.0"
click = "8.1.7"
colorama = "0.4.6"
coverage = "7.5.0"
dill = "0.3.8"
exceptiongroup = "1.2.1"
fastapi = "0.110.2"
frozenlist = "1.4.1"
h11 = "0.14.0"
httptools = "0.6.1"
idna = "3.7"
iniconfig = "2.0.0"
isort = "5.13.2"
mccabe = "0.7.0"
multidict = "6.0.5"
mypy = "1.10.0"
mypy-extensions = "1.0.0"
packaging = "24.0"
platformdirs = "4.2.1"
pluggy = "1.5.0"
pydantic = "2.7.1"
pydantic-core = "2.18.2"
pyenchant = "3.2.2"
pylint = "3.1.0"
pytest = "8.2.0"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
pyzmq = "26.0.2"
sniffio = "1.3.1"
starlette = "0.37.2"
tomli = "2.0.1"
tomlkit = "0.12.4"
typing-extensions = "4.11.0"
uvicorn = "0.29.0"
watchfiles = "0.21.0"
websockets = "12.0"
yarl = "1.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
